comments for clarity, removed unused imports

+++ dataset.py
@@ -1,7 +1,5 @@
 import torch
 import pandas as pd
-import config
-from tokenizer import BoardTokenizer
 import numpy as np
 
 class AnnotatedBoardsDataset(torch.utils.data.Dataset):
@@ -13,7 +11,8 @@ class AnnotatedBoardsDataset(torch.utils.data.Dataset):
     
     def __getitem__(self, idx):
         row = self.df.iloc[idx]
+        # ignore game id and cp evaluation
         _, fen, _, win_prob, draw_prob, lose_prob = tuple(row)
-        evaluation = np.array([win_prob, draw_prob, lose_prob], dtype=np.float32)
+        evaluation = np.array([win_prob, draw_prob, lose_prob], dtype=np.half)
         evaluation = torch.from_numpy(evaluation)
         return fen, evaluation
